Refactored users options in a CLI file
